Added input number widget
